+++ changelog.md
@@ -1,2 +1,7 @@
+* Added Mossy Planks, Stairs & Slabs
+* Added #blockus:all_wooden_mosaics & #blockus:all_mossy_planks block tags
+* Changed Birch Timber Frames textures
+* Removed old Bamboo Planks, Door & Trapdoor unused textures
 * Fixed [Rough red sandstone bug](https://github.com/Brandcraf06/Blockus/issues/320)
+* Fixed Herringbone Cherry Planks & Cherry Mosaic incorrect sounds